simplify multi engine abstraction. -4000 lines of code
