--- osbenchmark/database/clients/opensearch/__init__.py
@@ -1,11 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-#
-# The OpenSearch Contributors require contributions made to
-# this file be licensed under the Apache-2.0 license or a
-# compatible open source license.
-
-"""OpenSearch client implementation."""
-
-from osbenchmark.database.clients.opensearch.opensearch import OpenSearchClientFactory
-
-__all__ = ["OpenSearchClientFactory"]
