--- osbenchmark/database/clients/vespa/__init__.py
@@ -1,13 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-#
-# The OpenSearch Contributors require contributions made to
-# this file be licensed under the Apache-2.0 license or a
-# compatible open source license.
-
-from osbenchmark.database.clients.vespa.vespa import VespaClientFactory, VespaDatabaseClient, PYVESPA_AVAILABLE
-
-__all__ = [
-    'VespaClientFactory',
-    'VespaDatabaseClient',
-    'PYVESPA_AVAILABLE',
-]
